Revert "Refactor/render controller"
