+++ src/index.js
@@ -1,4 +1,3 @@
-import './js/ui-controller';
 import './js/preloader';
 import './js/render/render-gallery';
 import './js/render/render-library';
@@ -13,7 +12,8 @@ import './js/authentication';
 // import './js/empty-lib-modal';
 // import './js/empty-watch-modal';
 import './js/empty-lib-html';
-import './js/empty-watch-modal';
+
 import './js/modal';
+// import './js/storage/test';
 import './js/trailer';
 import './js/arrow';
